% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{add_label}
\alias{add_label}
\title{Add labels to reduced dimension plots}
\usage{
add_label(
  sce,
  dimname = "TSNE",
  text_by = "label",
  text_type = "text",
  text_size = 8,
  text_color = "black",
  box.padding = 0.5,
  max.overlaps = 20
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{dimname}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(sce)} to plot. Default is "TSNE".}

\item{text_by}{A string indicating the column metadata field with which to
add text labels on the plot. Alternatively, a character vector of the same
length as \code{colData(sce)} indicating the labels of each cell. Default is
"label".}

\item{text_type}{A string indicating to add text directly to the plot
(\code{"text"}) or draw a rectangle underneath the text (\code{"label"}). Default
is "text".}

\item{text_size}{A numeric scalar indicating the size of the label. This
is passed to \code{geom_text_repel()}. Default is 8.}

\item{text_color}{A string indicating the colour of the label. This is
passed to \code{geom_text_repel()}. Default is "black".}

\item{box.padding}{A scalar indicating the amount of padding around
bounding box, as unit or number. This is passed to \code{geom_text_repel()}.
Default is 0.5.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
This is passed to \code{geom_text_repel()}. Default is 20.}
}
\value{
A \code{geom} (geometric object)
}
\description{
This function is designed to be used with \code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}
from the \pkg{scater} package without specifying text label (i.e. via
\code{text_by}). Then use \code{add_label()} to place labels centrally.
}
\details{
The repel-away-from-center behaviour in \code{plotReducedDim()} should
be fixed in \pkg{scater} v1.23.5.
}
\examples{
library(scater)

# Load demo dataset
data(sce)

plotReducedDim(sce, "TSNE", colour_by = "label") + add_label(sce, "TSNE")
}
\seealso{
\code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}
}
\author{
I-Hsuan Lin
}
