% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUtils-package.R
\docType{data}
\name{data-dbl_results}
\alias{data-dbl_results}
\alias{dbl_results}
\title{findDoubletClusters results from a simulated single-cell RNA-seq dataset}
\format{
an object of class DFrame
}
\usage{
data(dbl_results)
}
\description{
findDoubletClusters results from a simulated single-cell RNA-seq dataset
}
\keyword{datasets}
