% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUtils-package.R
\docType{data}
\name{data-phases_assignments}
\alias{data-phases_assignments}
\alias{phases_assignments}
\title{cyclone results from a simulated single-cell RNA-seq dataset}
\format{
an object of class list
}
\usage{
data(phases_assignments)
}
\description{
cyclone results from a simulated single-cell RNA-seq dataset
}
\keyword{datasets}
