% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUtils-package.R
\docType{data}
\name{data-res_deseq2}
\alias{data-res_deseq2}
\alias{res_deseq2}
\title{A DESeq2 DESeqResults produced from the pasilla dataset}
\format{
an object of class DESeqResults
}
\usage{
data(res_deseq2)
}
\description{
A DESeq2 DESeqResults produced from the pasilla dataset
}
\keyword{datasets}
