% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUtils-package.R
\docType{data}
\name{data-res_edger}
\alias{data-res_edger}
\alias{res_edger}
\title{A edgeR TopTags object produced from the pasilla dataset}
\format{
an object of class TopTags
}
\usage{
data(res_edger)
}
\description{
A edgeR TopTags object produced from the pasilla dataset
}
\keyword{datasets}
