% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUtils-package.R
\docType{data}
\name{data-sce}
\alias{data-sce}
\alias{sce}
\title{A processed simulated single-cell RNA-seq dataset}
\format{
an object of class SingleCellExperiment
}
\usage{
data(sce)
}
\description{
A processed simulated single-cell RNA-seq dataset
}
\keyword{datasets}
