% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_basics.R
\name{discrete_colors}
\alias{discrete_colors}
\alias{c30}
\alias{c40}
\title{Discrete palettes}
\usage{
c30(n = 30)

c40(n = 40)
}
\arguments{
\item{n}{Number of colours to return. (default: maximum number of
available colours)}
}
\value{
A vector of colours as hex strings.
}
\description{
Colour palettes designed for discrete, categorical data with no more
than 40 categories.
}
\details{
The \code{c30()} palette has 30 unique colours and \code{c40()} palette has 40
unique colours. The \code{c40()} colour palette is taken from \code{plotScoreHeatmap()}
of the \pkg{SingleR} package (which itself is based on \pkg{DittoSeq} and
Okabe-Ito colors).
}
\examples{
# Show colours as pie charts
pie(rep(1, 30), col = c30(), main = "c30 palette")
pie(rep(1, 40), col = c40(), main = "c40 palette")
}
\seealso{
\code{\link[SingleR:plotScoreHeatmap]{SingleR::plotScoreHeatmap()}}
}
\author{
I-Hsuan Lin
}
