% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_markers.R
\name{exportResList}
\alias{exportResList}
\title{Export results from findMarkers, edgeR or DESeq2 to tsv}
\usage{
exportResList(
  object,
  concatenate = FALSE,
  col_anno = NULL,
  direction = "both",
  prefix = "output",
  dir_path = NULL
)
}
\arguments{
\item{object}{A named list of objects storing outputs returned by
\code{findMarkers()}, \pkg{edgeR} or \pkg{DESeq2}, each of which contains
results for the corresponding group of comparison.}

\item{concatenate}{Logical scalar indicating whether to concatenate
results stored in a list to a single file. Default is \code{FALSE}.}

\item{col_anno}{A character vector containing the names of the annotation
columns in the results stored in \code{object}, for example
\code{c("GeneID","Symbol")}. This is used to construct common annotation
column(s) when \code{concatenate = TRUE}. The row names of the result is used
to create a "Symbol" column in the concatenated output when
\code{col_anno = NULL}. Default is \code{NULL}.}

\item{direction}{A string indicating the directionality of expression
changes of markers/DEGs used to perform marker gene detection or
differential expression analysis. The string is added to the output file
name. Allowable character values are \code{"both"}, \code{"up"} and \code{"down"}.
Default  is \code{"both"}.}

\item{prefix}{A string indicating the prefix of output file. Default is
\code{"output"}.}

\item{dir_path}{The directory path for exported TSV. The file is exported
to the current working directory when \code{file_path = NULL}. Default is \code{NULL}.}
}
\description{
This function takes a list of objects containing results from
\code{findMarkers()}, \pkg{edgeR} or \pkg{DESeq2}, and save the content
to text file(s).
}
\details{
When \code{concatenate = TRUE}, the function concatenate the test
statistics from results stored in the list and outputs a single file.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(sce)

# Use result saved after running findMarkers and 'pval.type = "any"'
# Save results into individual files in the per-session temporary directory
exportResList(metadata(sce)[["findMarkers1"]], dir_path = tempdir())

# Save concatenate result in the per-session temporary directory
exportResList(metadata(sce)[["findMarkers4"]],
  concatenate = TRUE,
  prefix = "findMarkers4", dir_path = tempdir()
)
}
\author{
I-Hsuan Lin
}
