% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_basics.R
\name{fig}
\alias{fig}
\alias{reset.fig}
\title{Change \verb{repr.plot.*} behaviour}
\usage{
fig(
  width = 10,
  height = 7,
  pointsize = 12,
  bg = "white",
  antialias = "gray",
  res = 120,
  quality = 90,
  family = "sans"
)

reset.fig(
  width = 10,
  height = 7,
  pointsize = 12,
  bg = "white",
  antialias = "gray",
  res = 120,
  quality = 90,
  family = "sans"
)
}
\arguments{
\item{width}{Plotting area width in inches. Default is 10.}

\item{height}{Plotting area height in inches. Default is 7.}

\item{pointsize}{Text height in pt. Default is 12.}

\item{bg}{Background colour. Default is "white".}

\item{antialias}{Which kind of antialiasing to use for for lines and text?
'gray', 'subpixel' or 'none'? Default is "gray".}

\item{res}{PPI for rasterization. Defaults to 120.}

\item{quality}{Quality of JPEG format in \%. Default is 90.}

\item{family}{Font family. "sans", "serif", "mono" or a specific one.
Default is "sans".}
}
\description{
Change the behaviour of \verb{repr.plot.*}.
}
\details{
The function uses \code{\link[=options]{options()}} to change the behaviour of \verb{repr.plot.*}.
It provides a quick and easy way to change plot size and other
\verb{repr.plot.*} behaviours when running R in a Jupyter Notebook. If use
without indicating any argument, the plot behaviour will be reset to
default. The \code{reset.fig()} is an alias of \code{fig()}.
}
\examples{
\dontrun{
library(ggplot2)

# Change plot area width to 8 inches and height to 5 inches
fig(width = 8, height = 5)
ggplot(mpg, aes(class)) +
  geom_bar()

# Reset to default settings
fig()

# Change plot wider and taller
fig(width = 14, height = 10)
ggplot(mpg, aes(class)) +
  geom_bar()

# Alias of fig()
reset.fig()
}
}
\author{
I-Hsuan Lin
}
