% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_basics.R
\name{get_Legend}
\alias{get_Legend}
\title{Retrieve the legend of a plot}
\usage{
get_Legend(plot, legend = NULL)
}
\arguments{
\item{object}{A \code{ggplot} or \code{gtable} from which to retrieve the legend.}
}
\value{
A \code{gtable} object holding just the legend or \code{NULL} if there is no legend.
}
\description{
This function extracts just the legend from a ggplot.
}
\details{
The \code{get_legend} function from the cowplot package stopped working properly
with ggplot2 v3.5.0. A modified function was contributed by Teun van den Brand
(teunbrand) on cowplot's GitHub Issues and should work with the newer version of ggplot2.
Please refer to \url{https://github.com/wilkelab/cowplot/issues/202} on the discussion of
this bug and fixes.
}
\examples{
library(ggplot2)

p <- ggplot(mpg, aes(displ, hwy, colour = factor(cyl), shape = factor(year))) +
  geom_point() + guides(shape = guide_legend(position = "bottom"))

plot(get_Legend(p))
}
