% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_markers.R
\name{getDEGs}
\alias{getDEGs}
\title{Retrieve differentially expressed genes from edgeR or DESeq2 analysis}
\usage{
getDEGs(
  object,
  direction = "both",
  fdr = 0.05,
  logfc = c(0, 0),
  max = 1000,
  column_by = NULL
)
}
\arguments{
\item{object}{An object of class \code{TopTags}, \code{DGEExact}, \code{DGELRT} or
\code{DESeqResults}.}

\item{direction}{A string indicating the directionality of expression
changes of DEGs to retreive. Allowable character values are \code{"both"},
\code{"up"} and \code{"down"}. Default is "both".}

\item{fdr}{A numeric scalar indicating the FDR threashold used to select
DEGs. Default is 0.05.}

\item{logfc}{A numeric scalar indicating the logFC cutoffs used to select
DEGs. Default is \code{c(0, 0)}.}

\item{max}{An integer scalar indicating the maximum number of DEGs to
return. Default is 1000.}

\item{column_by}{A string indicating the column name containing the gene
symbols or IDs in the input \code{object} and the selected information is
returned. When \code{column_by = NULL}, it retreives the row names from the
DataFrame. Default is \code{NULL}.}
}
\value{
A character vector
}
\description{
This function parses an object of class \code{TopTags}, \code{DGEExact} or \code{DGELRT}
from \pkg{edgeR} or \code{DESeqResults} class from \pkg{DESeq2}, and returns
the differentially expressed genes (DEGs) that passed the specified
thresholds.
}
\details{
The function parses the standard output objects from \pkg{edgeR} and
\pkg{DESeq2} to selects DEGs that "passed" the \code{fdr}, \code{logfc} and \code{max}
threshold.

The accapted class types includes:
\itemize{
\item \code{DGEExact}, returned by \code{exactTest()} from \pkg{edgeR}.
\item \code{DGELRT}, returned by \code{glmLRT()}, \code{glmTreat()} or \code{glmQLFTest()}
from \pkg{edgeR}.
\item \code{TopTags}, returned by \code{topTags()} from \pkg{edgeR}.
\item \code{DESeqResults} returned by \code{results()} from \pkg{DESeq2}.
}

By specifying \code{direction}, the function selects DEGs with a positive and
negative logFC when \code{direction = "both"} (default), only positive logFC when
\code{direction = "up"}, and only negative logFC when \code{direction = "down"}.
}
\examples{
# Load demo dataset
data(res_deseq2)
data(res_edger)

# Retrieve first 250 up-regulated DEGs from DESeq2's output,
# return rownames (Ensembl ID)
getDEGs(res_deseq2, direction = "up", max = 250)

# Retrieve first 1000 DEGs at FDR = 10\% from edgeR's output,
# return values from the `external_gene_name` column (Gene Symbol)
getDEGs(res_edger, fdr = 0.1, column_by = "external_gene_name")
}
\seealso{
\code{\link[edgeR:exactTest]{edgeR::exactTest()}}, \code{\link[edgeR:glmLRT]{edgeR::glmLRT()}}, \code{\link[edgeR:glmTreat]{edgeR::glmTreat()}},
\code{\link[edgeR:glmQLFTest]{edgeR::glmQLFTest()}}, \code{\link[edgeR:topTags]{edgeR::topTags()}}, \code{\link[DESeq2:results]{DESeq2::results()}}
}
\author{
I-Hsuan Lin
}
