% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_markers.R
\name{getMarkers}
\alias{getMarkers}
\title{Retrieve marker genes from a \code{findMarkers} output DataFrame}
\usage{
getMarkers(
  object,
  direction = "both",
  fdr = 0.05,
  top = 200,
  logfc = c(0, 0),
  auc = c(0.3, 0.7),
  max = 1000,
  column_by = NULL
)
}
\arguments{
\item{object}{A DataFrame that contain a sorted marker gene list for a
single group. It should be taken from a named list of DataFrames returned
by \code{findMarkers()}.}

\item{direction}{A string indicating the directionality of expression
changes of markers to retreive. Allowable character values are \code{"both"},
\code{"up"} and \code{"down"}. Default is "both".}

\item{fdr}{A numeric scalar indicating the FDR threashold used to select
markers. This threshold has no effect when the input source is from
\code{findMarkers()} ran with \code{pval.type = "any"}. Default is 0.05.}

\item{top}{An integer scalar indicating the 'Top' threashold used to select
markers. This threshold is only applicable when the input source is from
\code{findMarkers()} ran with \code{pval.type = "any"}. Default is 200.}

\item{logfc}{A numeric vector of length 2 specifying the logFC threasholds
used to select up-/down-regulated markers/DEGs. This threshold has
no effect when the input source is from \code{findMarkers()} ran with
\code{test.type = "wilcox"}. Default is \code{c(0, 0)}.}

\item{auc}{A numeric vector of length 2 specifying the AUC threasholds
used to select up-/down-regulated markers. This threshold is only
applicable when the input source is from \code{findMarkers()} ran with
\code{test.type = "wilcox"}. Default is \code{c(0.3, 0.7)}.}

\item{max}{An integer scalar indicating the maximum number of markers to
return. Default is 1000.}

\item{column_by}{A string indicating the column name containing the gene
symbols or IDs in the input \code{object} and the selected information is
returned. When \code{column_by = NULL}, it retreives the row names from the
DataFrame. Default is \code{NULL}.}
}
\value{
A character vector
}
\description{
This function parses a \code{findMarkers()} output DataFrame and returns the
up- and down-regulated marker genes that passed the specified threshold.
}
\details{
The function takes a DataFrame from the \code{findMarkers()} output (which is
a named list of DataFrames) and returns the marker genes that "passed" the
\code{fdr} or \code{top} threshold.

By specifying \code{direction}, the function selects up-regulated markers when
\code{direction = "up"}, down-regulated markers when \code{direction = "down"} and
both up- and down-regulated markers when \code{direction = "both"} (default).
The definition of up/down-regulation depends on the \code{test.type} used when
running \code{findMarkers()}.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(sce)

# Retrieve first 1000 markers (default `max`) from the 1st DataFrame
# in 'findMarkers1', return rownames
getMarkers(metadata(sce)[["findMarkers1"]][[1]])

# Retrieve first 1000 up-regulated markers from the 2nd DataFrame
# in 'findMarkers4', return rownames
getMarkers(metadata(sce)[["findMarkers4"]][[2]], direction = "up")
}
\seealso{
\code{\link[scran:findMarkers]{scran::findMarkers()}}
}
\author{
I-Hsuan Lin
}
