% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_net.R
\name{netVis_heatmap}
\alias{netVis_heatmap}
\title{Visualization of network using heatmap}
\usage{
netVis_heatmap(
  obj,
  comparison = c(1, 2),
  measure = c("count", "weight", "prob"),
  pathway = NULL,
  heatmap_colors = c("#4575b4", "#d73027"),
  heatmap_colors_min = NULL,
  heatmap_colors_max = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  row_names_rot = 0,
  column_names_rot = 90,
  fontsize = 9,
  cell_fontsize = 9,
  show_values = TRUE,
  legend_side = "bottom",
  legend_direction = "horizontal",
  legend_title = NULL,
  legend_title_position = "leftcenter",
  legend_width = unit(5, "cm"),
  legend_height = NULL,
  right_anno_width = unit(2, "cm"),
  top_anno_height = unit(2, "cm"),
  anno_colors = c("#66c2a5", "#fc8d62"),
  anno_legend_ncol = 1,
  anno_legend_side = "bottom",
  anno_legend_direction = "horizontal",
  anno_legend_title = "Sample",
  anno_legend_title_position = "leftcenter",
  plot_title = NULL,
  title_prefix = NULL,
  row_title = "Sources (outgoing signaling)",
  draw = TRUE,
  data_only = FALSE
)
}
\arguments{
\item{obj}{A CellChat object containing one or merged datasets.}

\item{comparison}{A numerical or character vector of length 2, to indicate the
datasets for comparison. Default is \code{c(1, 2)}.}

\item{measure}{Define the measure to plot, \code{count} shows the number of interactions;
\code{weight} shows the total interaction weights (strength);
\code{prob} shows thecommunication probabilities of a chosen pathway.}

\item{pathway}{A string to indicate the name of signaling networks in the CellChat
object. This is required when \code{measure = "prob"}.}

\item{heatmap_colors}{A character vector to indicate the colours used to generate a
colour palette. The first colour provided is used to create a sequential palette
for single-data CellChat object. When comparing two datasets, a diverging palette
is created to match the max/min values. Default is \code{c("#4575b4", "#d73027")}.}

\item{heatmap_colors_min}{A numeric value to set the minimum value used to generate
the colour palette. Default is \code{NULL} and is calculated from the data.}

\item{heatmap_colors_max}{A numeric value to set the maximum value used to generate
the colour palette. Default is \code{NULL} and is calculated from the data.}

\item{cluster_rows}{A logical value to indicate whether to cluster on rows.
Default is \code{FALSE}.}

\item{cluster_columns}{A logical value to indicate whether to cluster on columns.
Default is \code{FALSE}.}

\item{row_names_rot}{A numeric value to set the rotation of row names.
Default is \code{0}.}

\item{column_names_rot}{A numeric value to set the rotation of column names.
Default is \code{90}.}

\item{fontsize}{A numeric value to set the overall font size. Default is \code{9}.}

\item{cell_fontsize}{A numeric value to set the font size of the values in each
cell. Default is \code{9}.}

\item{show_values}{A logical value to indicate whether to show the values in each
cell. Default is \code{TRUE}.}

\item{legend_side}{A string to indicate the side to put heatmap legend.
Possible options are \code{right}, \code{left}, \code{bottom}, and \code{top}. Default is \code{bottom}.}

\item{legend_direction}{A string to indicate the direction of the heatmap legend.
Possible options are \code{vertical} and \code{horizontal}. Default is \code{horizontal}.}

\item{legend_title}{A string to indicate the heatmap legend title. Default is \code{NULL}
and the function will generate a suitable title depending on the selected \code{measure}.}

\item{legend_title_position}{A string to indicate the position of title relative
to the heatmap legend. Possible options are \code{leftcenter} and \code{lefttop} for horizontal
legend. Vertical legend accepts \code{topleft}, \code{topcenter}, \code{leftcenter-rot} and
\code{lefttop-rot}. Default is \code{leftcenter}.}

\item{legend_width}{A \code{unit} object to set the width of the the whole heatmap legend
body. It is only used for horizontal continous legend. Default is \code{unit(5, "cm")}.}

\item{legend_height}{A \code{unit} object to set the height of the whole heatmap legend
body. It is only used for vertical continous legend. Default is \code{NULL}.}

\item{right_anno_width}{A \code{unit} object to set the width of the right (row)
annotation on the right. Default is \code{unit(2, "cm")}.}

\item{top_anno_height}{A \code{unit} object to set the height of the top (column)
annotations on the top. Default is \code{unit(2, "cm")}.}

\item{anno_colors}{A character vector to indicate the colours used to fill the
box plots showing the two datasets, or a single colour used to fill the bar plots
showing a single dataset. Default is \code{c("#66c2a5", "#fc8d62")}.}

\item{anno_legend_ncol}{A integer to set the number of columns in the legend grids.
Default is \code{1}.}

\item{anno_legend_side}{A string to indicate the side to put annotation legend.
Default is \code{bottom}.}

\item{anno_legend_direction}{A string to indicate the direction of the annotation
legend. Default is \code{horizontal}.}

\item{anno_legend_title}{A string to indicate the annotation legend title.
Default is "Sample".}

\item{anno_legend_title_position}{A string to indicate the position of title relative
to the annotation legend. Default is \code{leftcenter}.}

\item{plot_title}{A string to indicate the plot (column) title. Default is \code{NULL}
and the function will generate a suitable title depending on the selected \code{measure}.}

\item{title_prefix}{A string to append to the plot (column) title. Default is \code{NULL}.}

\item{row_title}{A string to indicate the row title.
Default is \verb{Sources (outgoing signaling)}.}

\item{draw}{A logical value to indicate whether to draw the heatmap.
When \code{draw = FALSE}, the function retuns a grob (\code{gTree}) object made with package
\code{grid}. Default is \code{TRUE}.}

\item{data_only}{A logical value to indicate whether to return a list containing the
calculated values. This is useful for users to create their own plots.
When \code{data_only = TRUE}, the function will not plot or return a \code{gTree} object.
Default is \code{FALSE}.}
}
\value{
A plot appears on currect plotting device if \code{draw = FALSE}.
When \code{draw = FALSE}, the function retuns a grob (\code{gTree}) object made with package
\code{grid}. When \code{data_only = TRUE}, the function retuns \code{list} object containing the
data values.
}
\description{
This heatmap can be used to show number of interactions, interaction strength,
communication probabilities of a chosen pathway in a single dataset
(defined by \verb{pathway =}), as well as the differential comparison of these
three measures between two datasets.
}
\details{
When showing differential comparison:
\itemize{
\item Using the default \code{heatmap_colors}, the red (or blue) represents increased
(or decreased) signaling in the second dataset compared to the first one.
\item The box plots on the top show the original values of the selected \code{measure}
from cell groups from the two datasets in each column (incoming signaling).
\item The box plots on the right show the original values of the selected \code{measure}
from cell groups from the two datasets in each row (outgoing signaling).
}

The function offers an alternative method to create heatmap from CellChat object.

Compared to \code{CellChat::netVisual_heatmap} function, it offers more controls to the
plot aesthetic and offers two output types (defined by \verb{draw =}): to plot on device
or provide a \code{gTree} object. This can be used with other R package such as 'patchwork'
(might require warpping \code{gTree} object with \code{wrap_elements} function), 'cowplot'
(with \code{plot_grid} function) or 'gridExtra' (with \code{grid.arrange} function) to combine
multiple heatmaps, or with other type of plot object types such as \code{ggplot} into a
single graphic.

For users who will like to create their own plots, one can use \code{data_only = TRUE} to
obtain a list that contains the matrix with the calculated values, and optionally the
original values of the selected \code{measure} from cell groups from the two datasets when
showing differential comparison.
}
\author{
I-Hsuan Lin
}
