% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotCyclone}
\alias{plotCyclone}
\title{Plot G2M against G1 scores from \code{cyclone} results}
\usage{
plotCyclone(
  x,
  phase_color = NULL,
  point_size = 2,
  point_alpha = 0.8,
  text_size = 6,
  xlab = "G1 score",
  ylab = "G2/M score",
  title = NULL,
  theme_size = 18
)
}
\arguments{
\item{x}{A \code{list} returned by \code{cyclone()} containing \code{phases}, \code{scores}
and \code{normalized.scores} slots.}

\item{phase_color}{A character vector of color codes indicating the colour
of the cell cycle phases. Default is \code{NULL}, and use \code{\link[=choosePalette]{choosePalette()}} to
select a palette.}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{point_alpha}{A numeric scalar (between 0 and 1) indicating the
transparency. Default is 0.8.}

\item{text_size}{A numeric scalar indicating the size of cell cycle phase
labels. Default is 6.}

\item{xlab}{The title of the x-axis. Default is "G1 score".}

\item{ylab}{The title of the y-axis. Default is "G2/M score".}

\item{title}{Plot title. Default is \code{NULL}.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}
}
\value{
A \code{ggplot} object
}
\description{
This function produces a scatter plot to show the predicted G1 and
G2M scores and cell cycle phases of single cells after performing
classification using \code{cyclone()} from the \pkg{scran} package.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(phases_assignments)

# Create plot, minimal input
plotCyclone(phases_assignments)

# Assign colours
my.cols <- c("red", "blue", "black")
plotCyclone(phases_assignments, phase_color = my.cols)

# Assign named colours
my.cols <- setNames(c("red", "blue", "black"), c("G1", "S", "G2M"))
plotCyclone(phases_assignments, phase_color = my.cols)
}
\seealso{
\code{\link[scran:cyclone]{scran::cyclone()}}
}
\author{
I-Hsuan Lin
}
