% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotProjection}
\alias{plotProjection}
\title{Colour cells by a feature in a 2-dimension representation}
\usage{
plotProjection(
  sce,
  feature,
  dimname = "TSNE",
  feat_desc = NULL,
  feat_color = NULL,
  color_breaks = waiver(),
  exprs_by = "logcounts",
  text_by = NULL,
  point_size = 2,
  point_alpha = 0.5,
  theme_size = 18,
  legend_pos = "right",
  legend_just = "center",
  guides_ncol = NULL,
  guides_nrow = NULL,
  guides_barwidth = NULL,
  guides_barheight = NULL,
  guides_size = point_size * 2,
  title = NULL,
  show_title = TRUE,
  show_subtitle = TRUE,
  other_fields = list(),
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{feature}{A string indicating the column name in \code{colData(sce)}
containing the feature to be coloured. Alternatively, a character vector of
the same length as \code{colData(sce)} indicating the feature type of each cell.}

\item{dimname}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(sce)} to plot. Default is "TSNE".}

\item{feat_desc}{A string that describes the coloured feature. Default is
\code{NULL}.}

\item{feat_color}{A character vector of colour codes indicating the colours
of the features, or a palette function that creates a vector of colours
along a colour map. Default is \code{NULL}, and use \code{\link[=choosePalette]{choosePalette()}} to
select a palette.}

\item{color_breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}

\item{exprs_by}{A string or integer scalar specifying which assay from
\code{sce} to obtain expression values from, for use in point aesthetics. Use
\code{assayNames(sce)} to find all availavle assays in \code{sce}. Default is
"logcounts".}

\item{text_by}{A string indicating the column metadata field with which
to add text labels on the plot. Default is \code{NULL}.}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{point_alpha}{A numeric scalar (between 0 and 1) indicating the
transparency. Default is 0.5.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}

\item{legend_pos}{The position of legends ("none", "left", "right",
"bottom","top". Use "none" to disable plot legend. Default is "right".}

\item{legend_just}{The anchor point for positioning legend inside plot
("center" or two-element numeric vector) or the justification according
to the plot area when positioned outside the plot. Default is "center".}

\item{guides_ncol, guides_nrow}{An integer scalar indicating the desired
number of column and row of discrete legends. Default is NULL.}

\item{guides_barwidth, guides_barheight}{A numeric or a \code{grid::unit()}
object specifying the width and height of the colourbar. Default is NULL.}

\item{guides_size}{An integer scalar indicating the desired size of the
points in a discrete legend. Default is \code{point_size * 2}.}

\item{title}{Plot title. Default is \code{NULL} and the \code{dimname} is used as
title.}

\item{show_title}{Logical scalar indicating whether to show plot title.
Default is \code{TRUE}.}

\item{show_subtitle}{Logical scalar indicating whether to show plot
subtitle, Default is \code{TRUE}.}

\item{other_fields}{Additional cell-based fields to include in the
DataFrame. Default is \code{list()}.}

\item{...}{Other arguments passed on to \code{\link[=add_label]{add_label()}}.}
}
\value{
A \code{ggplot} object
}
\description{
This function uses \code{plotReducedDim()} from the \pkg{scater} package to
show cells on a pre-calculated low-dimensional projection (such as UMAP or
t-SNE) and colour by a choosen cell-specific feature. It uses \code{add_label()}
to label cells when \code{text_by} is used.
}
\details{
The function uses \code{plotReducedDim()} from the \pkg{scater} package to
create a base plot and adds aesthetic, such as point colours, legend
controls, title and subtitles, to the final figure.
}
\examples{
# Load demo dataset
data(sce)

# Colour and label cells by cluster
plotProjection(sce, "label",
  dimname = "TSNE", text_by = "label",
  feat_desc = "Cluster"
)

# Colour cells by DKD62 expression and label by cluster
plotProjection(sce, "DKD62",
  dimname = "UMAP", text_by = "label",
  feat_desc = "DKD62 Expression", guides_barheight = 15
)
}
\seealso{
\code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}
}
\author{
I-Hsuan Lin
}
