% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotqcDoubletClusters}
\alias{plotqcDoubletClusters}
\title{Create QC plots from \code{findDoubletClusters} output}
\usage{
plotqcDoubletClusters(
  dbl,
  clusters = NULL,
  cluster_color = NULL,
  qc_plot = 0,
  point_size = 2,
  text_size = 6,
  box.padding = 0.5,
  point.padding = 0.5,
  max.overlaps = Inf,
  seed = 12321,
  theme_size = 18
)
}
\arguments{
\item{dbl}{A DataFrame returned by \code{findDoubletClusters()} that identifies
potential clusters of doublet cells.}

\item{clusters}{A vector of length equal to number of rows in \code{dbl},
indicating the cluster levels. Default is \code{NULL} and use the row names in
\code{dbl}.}

\item{cluster_color}{A character vector of color codes indicating the
colour of the clusters. Default is \code{NULL}, and use \code{\link[=choosePalette]{choosePalette()}} to
select a palette.}

\item{qc_plot}{A numeric scalar indicating which QC plot to create.
Default is 0 to produce a compound figure, or:
\itemize{
\item 1: A scatter plot of median number of significant genes (\code{median.de})
against number of significant genes (\code{num.de}).
\item 2: A barplot showing the proportion of cells in each of the query
cluster (\code{prop}).
\item 3: A scatter plot of ratio of the median library sizes for the second
source cluster (\code{lib.size2}) against first source cluster (\code{lib.size1}).
}}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{text_size}{A numeric scalar indicating the size of the label.
This is passed to \code{geom_text_repel()}. Default is 8.}

\item{box.padding}{A scalar indicating the amount of padding around
bounding box, as unit or number. This is passed to \code{geom_text_repel()}.
Default is 0.5.}

\item{point.padding}{A scalar indicating the amount of padding around
labeled point, as unit or number. This is passed to \code{geom_text_repel()}.
Default is 0.5.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
This is passed to \code{geom_text_repel()}. Default is Inf.}

\item{seed}{Random seed passed to \code{\link[=set.seed]{set.seed()}}. This is passed to
\code{geom_text_repel()}. Default is 12321.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}
}
\value{
A \code{ggplot} object when \code{qc_plot} is not 0, or a \pkg{ggplot2}
plot with an object of class \code{c("gg", "ggplot")} when \code{qc_plot = 0}
}
\description{
This function produces QC plots to aid deciding potential doublet clusters
after running \code{findDoubletClusters()} from the \pkg{scDblFinder} package.
}
\examples{
library(SingleCellExperiment)

data(dbl_results)

# Create compound figure
plotqcDoubletClusters(dbl_results)

# Create median.de against num.de
plotqcDoubletClusters(dbl_results, qc_plot = 1)
}
\seealso{
\code{\link[scDblFinder:findDoubletClusters]{scDblFinder::findDoubletClusters()}}
}
\author{
I-Hsuan Lin
}
