% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot distribution of approximate silhouette widths}
\usage{
plotSilhouette(
  object,
  clusters,
  printDiff = TRUE,
  plot = TRUE,
  cluster_color = NULL,
  point_size = 2,
  point_alpha = 0.8,
  point_shape = 16,
  swarm_method = "quasirandom",
  add_mean = TRUE,
  mean_color = "firebrick",
  mean_size = 1,
  theme_size = 18
)
}
\arguments{
\item{object}{A numeric matrix-like object containing observations in rows
and variables in columns.}

\item{clusters}{A vector of length equal to \code{ncol(object)}, indicating the
cluster assigned to each observation.}

\item{printDiff}{Logical scalar indicating whether to print the precentage
of cells assigned to different cluster(s). Default is \code{TRUE}.}

\item{plot}{Logical scalar indicating whether to draw the plot. Default is
\code{TRUE}.}

\item{cluster_color}{A character vector of color codes indicating the
colour of the clusters. Default is \code{NULL}, and use \code{\link[=choosePalette]{choosePalette()}} to
select a palette.}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{point_alpha}{A numeric scalar (between 0 and 1) indicating the
transparency. Default is 0.8.}

\item{point_shape}{An integer scalar (between 0 and 25) indicating the
shape aesthetics. Default is 16.}

\item{swarm_method}{A string indicating the method for arranging points.
Method recongised by \pkg{ggbeeswarm} are "quasirandom", "pseudorandom",
"smiley" and "frowney". Default is "quasirandom".}

\item{add_mean}{Logical scalar indicating whether add a horizontal
reference line at mean silhouette width. Default is \code{TRUE}.}

\item{mean_color}{A string indicating the colour of the reference line.
Default is "firebrick".}

\item{mean_size}{A numeric scalar indicating the size of the reference
line. Default is 1.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}
}
\value{
A \code{ggplot} object when \code{plot = TRUE}
}
\description{
This function creates a violin scatter plot to show the distribution of
the approximate silhouette width across cells in each cluster.
}
\details{
The function uses \code{approxSilhouette()} from the \pkg{bluster} package' to
calculate the silhouette widths using an approximate approach. Using the
information stored in the returned DataFrame, it:
\itemize{
\item prints the descriptive statistics of approximate silhouette widths.
\item calculates the percentage of cells assigned to the closest cluster and
prints the result when \code{printDiff = TRUE}.
\item creates a violin scatter plot to show thedistribution of the approximate
silhouette width when \code{plot = TRUE}.
}

These information can allow users to identify poorly separate clusters
quickly.
}
\examples{
# Load demo dataset
data(sce)

mat <- SingleCellExperiment::reducedDim(sce, "PCA")
plotSilhouette(mat, sce$label)
}
\seealso{
\code{\link[bluster:approxSilhouette]{bluster::approxSilhouette()}}
}
\author{
I-Hsuan Lin
}
