% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_markers.R
\name{printEnrichR}
\alias{printEnrichR}
\title{Export results from Enrichr to tsv}
\usage{
printEnrichR(
  object,
  prefix = "enrichr",
  showTerms = NULL,
  columns = c(1:9),
  dir_path = NULL
)
}
\arguments{
\item{object}{A named list of list of DataFrames storing outputs returned
by \code{runEnrichR()}, each of which contains results for the corresponding
group of comparison.}

\item{prefix}{A string indicating the prefix of output file.
Default is "enrichr".}

\item{showTerms}{An integer scalar indicating the number of terms to save
to file. All terms are saved when \code{showTerms = NULL}. Default is \code{NULL}.}

\item{columns}{An integer vector indicating the columns from each entry
of data to save to file. All columns are saved when \code{columns = c(1:9)}.
Default is \code{c(1:9)}.
1-"Term", 2-"Overlap", 3-"P.value", 4-"Adjusted.P.value",
5-"Old.P.value", 6-"Old.Adjusted.P.value", 7-"Odds.Ratio",
8-"Combined.Score",  9-"Combined.Score".}

\item{dir_path}{The directory path for exported TSV. The files are exported
to the current working directory when \code{file_path = NULL}. Default is \code{NULL}.}
}
\description{
This function takes a list of list of DataFrames containing Enrichr results
returned by \code{runEnrichR()} and save all the results in the list object to
individual text files.
}
\details{
We are not using \code{printEnrich()} in the \pkg{enrichR} package to
print results due to a bug not yet fixed in the current verion in CRAN at
the time of writing (v3.0).
}
\examples{
\dontrun{
# Load demo dataset
data(res_deseq2)

# We construct a list of 3 DataFrames using the same DESeq2 output (`res_deseq2`)
# to mimic having a list containing results from multiple sets of comparisons.
res.de <- list(A_B = res_deseq2, A_C = res_deseq2, B_C = res_deseq2)

# Select gene-set libraries
dbs <- c("GO_Molecular_Function_2018", "GO_Biological_Process_2018")

# Run enrichR using the D. melanogaster specific modEnrichr site, and
# specify the gene symbols are stored in the 'external_gene_name' column
res.ora <- runEnrichR(res.de,
  dbs = dbs, site = "FlyEnrichr",
  column_by = "external_gene_name"
)

# Save results to TSV files in the per-session temporary directory
printEnrichR(res.ora, prefix = "Enrichr", dir_path = tempdir())
}
}
\seealso{
\code{\link[=runEnrichR]{runEnrichR()}}
}
\author{
I-Hsuan Lin
}
