% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_markers.R
\name{printMarkerStats}
\alias{printMarkerStats}
\title{Print marker gene stats from \code{findMarkers} output}
\usage{
printMarkerStats(
  object,
  fdr = 0.05,
  top = 200,
  pval.type = NULL,
  min.prop = NULL,
  logfc = c(0, 0),
  auc = c(0.3, 0.7)
)
}
\arguments{
\item{object}{A named list of DataFrames returned by \code{findMarkers()}, each
of which contains a sorted marker gene list for the corresponding group of
comparison.}

\item{fdr}{A numeric scalar indicating the FDR threashold used to select
markers. This threshold has no effect when the input source is from
\code{findMarkers()} ran with \code{pval.type = "any"}. Default is 0.05.}

\item{top}{An integer scalar indicating the 'Top' threashold used to select
markers. This threshold is only applicable when the input source is from
\code{findMarkers()} ran with \code{pval.type = "any"}. Default is 200.}

\item{pval.type}{A string specifying how p-values are to be combined across
pairwise comparisons for a given group/cluster when using \code{findMarkers()}.
Default is \code{NULL}.}

\item{min.prop}{A numeric scalar specifying the minimum proportion of
significant comparisons per gene when using \code{findMarkers()}. Default is
\code{NULL}.}

\item{logfc}{A numeric vector of length 2 specifying the logFC threasholds
used to select up-/down-regulated markers/DEGs. This threshold has
no effect when the input source is from \code{findMarkers()} ran with
\code{test.type = "wilcox"}. Default is \code{c(0, 0)}.}

\item{auc}{A numeric vector of length 2 specifying the AUC threasholds
used to select up-/down-regulated markers. This threshold is only
applicable when the input source is from \code{findMarkers()} ran with
\code{test.type = "wilcox"}. Default is \code{c(0.3, 0.7)}.}
}
\value{
‘simple’ diagnostic messages
}
\description{
This function parses the \code{findMarkers()} output (from the \verb{pkg\{scran\} package) and print the number of marker genes that "passed" the }fdr\code{or}top` threshold on to standard output in interactive sessions. It
provides a quick way to find out the number of candidate marker genes.
}
\details{
The function can detect the \code{pval.type} and \code{test.type}, to a certain
degree, based on the resulting DataFrame structure. One can optionally
provide the \code{pval.type} and \code{min.prop} settings used to run \code{findMarkers()}
to allow \code{printMarkerStats()} to use them in the printed messages.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(sce)

# Use result saved after running `findMarkers()` with `pval.type = "any"`
printMarkerStats(metadata(sce)[["findMarkers1"]])

# Use result saved after running `findMarkers()` with `pval.type = "all"`
printMarkerStats(metadata(sce)[["findMarkers2"]])

# Use result saved after running `findMarkers()` with `pval.type = "any"`
# and `min.prop = 0.5`
printMarkerStats(metadata(sce)[["findMarkers3"]], min.prop = 0.5)

# Use result saved after running `findMarkers()` with `test.type = "wilcox"`
printMarkerStats(metadata(sce)[["findMarkers4"]])
}
\seealso{
\code{\link[scran:findMarkers]{scran::findMarkers()}}
}
\author{
I-Hsuan Lin
}
