% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_basics.R
\name{choosePalette}
\alias{choosePalette}
\title{Choose discrete colours}
\usage{
choosePalette(object, color = c30(), quiet = TRUE)
}
\arguments{
\item{object}{A character vector containing the discrete or categorical
features.}

\item{color}{A character vector of color codes. Default is to use \code{c30()},
a palette with 30 colours.}

\item{quiet}{Logical scalar indicating whether to disable messaging.
Default is \code{TRUE}.}
}
\value{
A named character vector
}
\description{
Given a character vector of features and optionally a vector of color codes,
this function evaluate if the supplied color codes has sufficient number
of colours. It returns a named vector of color codes based on the input
features, with the same length as the unique features.
}
\details{
By default, it uses the \code{\link[=c30]{c30()}} palette when no more than 30 colours are
required, then the \code{\link[=c40]{c40()}} palette, and lastly the \code{\link[=rainbow]{rainbow()}} colour palette
when requiring more than 40 colours.
}
\examples{
# When input is a vector
feat <- LETTERS[1:5]
feat
choosePalette(feat) # use c30()

# When input is a factor of 3 levels
feat <- factor(rep(LETTERS[1:3], 5))
feat
choosePalette(feat, rainbow(10))
}
\seealso{
\code{\link[=c30]{c30()}}, \code{\link[=c40]{c40()}}, \code{\link[grDevices:palettes]{grDevices::rainbow()}}
}
\author{
I-Hsuan Lin
}
