% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_basics.R
\name{geom_parallel_sets_labs}
\alias{geom_parallel_sets_labs}
\title{Add nudged labels in a parallel sets diagram}
\usage{
geom_parallel_sets_labs(
  mapping = NULL,
  data = NULL,
  stat = "parallel_sets_axes",
  position = "identity",
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{A string indicating the statistical transformation to use on
the data for this layer. Default is "parallel_sets_axes".}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function. Cannot be jointy specified with
\code{nudge_x} or \code{nudge_y}.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{\link[grDevices:plotmath]{grDevices::plotmath()}}. Default is \code{FALSE}.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels
by. Useful for offsetting text from points, particularly on discrete
scales. Cannot be jointly specified with \code{position}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{layer()}.}
}
\value{
A \code{layer} object
}
\description{
This is the same function as \code{geom_parallel_sets_labels()} from the
\pkg{ggforce} package but with the ability to nudge labels at a fixed
distance. It is especially useful when the labels are too long to fit
inside the bars depicting the discrete categories.
}
\details{
A pull request of the nudge enhancement has been
submitted to the \pkg{ggforce}'s GitHub repository,
\url{https://github.com/thomasp85/ggforce/pull/260}, added in v0.3.4.
}
\examples{
library(ggforce)
data <- as.data.frame(Titanic)
data <- gather_set_data(data, 1:4)

# Use nudge_x to offset and hjust = 0 to left-justify label
ggplot(data, aes(x, id = id, split = y, value = Freq)) +
  geom_parallel_sets(aes(fill = Sex), alpha = 0.3, axis.width = 0.1) +
  geom_parallel_sets_axes(axis.width = 0.1) +
  geom_parallel_sets_labs(colour = "red", angle = 0, nudge_x = 0.1, hjust = 0)
}
\seealso{
\code{\link[ggforce:geom_parallel_sets]{ggforce::geom_parallel_sets_labels()}}, \code{\link[ggplot2:layer]{ggplot2::layer()}}
}
\author{
I-Hsuan Lin
}
