% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotExprsFreqVsMean}
\alias{plotExprsFreqVsMean}
\title{Plot expression frequency against mean counts}
\usage{
plotExprsFreqVsMean(
  sce,
  point_size = 2,
  point_alpha = 0.8,
  anno_size = 6,
  text_size = 4,
  text_color = "black",
  trend_color = "firebrick",
  trend_size = 1,
  trend_se = TRUE,
  box.padding = 0.5,
  max.overlaps = Inf,
  seed = 12321,
  xlab = "log2 Mean expression",
  ylab = "Percentage of expressing cells",
  title = NULL,
  theme_size = 18,
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing the required QC
metrics for features ('mean' and 'detected') that were returned by
\code{addPerCellQC()} or \code{perFeatureQCMetrics()} in its \code{rowData} slot.}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{point_alpha}{A numeric scalar (between 0 and 1) indicating the
transparency. Default is 0.8.}

\item{anno_size}{A numeric scalar indicating the size of annotation text.
Default is 6.}

\item{text_size}{A numeric scalar indicating the size of the label. This
is passed to \code{geom_text_repel()}. Default is 4.}

\item{text_color}{A string indicating the colour of the label. This is
passed to \code{geom_text_repel()}. Default is "black".}

\item{trend_color}{A string indicating the colour of the smoothed curve.
Default is "firebrick",}

\item{trend_size}{A numeric scalar indicating the size of the smoothed
curve. Default is 1.}

\item{trend_se}{Logical scalar indicating whether to display confidence
interval around smooth. Default is \code{TRUE}.}

\item{box.padding}{A scalar indicating the amount of padding around
bounding box, as unit or number. This is passed to \code{geom_text_repel()}.
Default is 0.5.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
This is passed to \code{geom_text_repel()}. Default is Inf.}

\item{seed}{Random seed passed to \code{\link[=set.seed]{set.seed()}}. This is passed to
\code{geom_text_repel()}. Default is 12321.}

\item{xlab}{The title of the x-axis. Default is "log2 Mean expression".}

\item{ylab}{The title of the y-axis. Default is "Percentage of expressing
cells".}

\item{title}{Plot title. Default is \code{NULL}.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}

\item{...}{Other arguments passed on to \code{geom_text_repel()}.}
}
\value{
A \code{ggplot} object
}
\description{
This function uses \code{plotRowData()} from the \pkg{scater} package to
produce a scatter plot of expression frequency against mean counts by
using per-feature quality control metrics produced by \pkg{scuttle}.
}
\details{
The original function of the same name can be found in the legacy
\pkg{scater} package and is now deprecated.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(sce)

plotExprsFreqVsMean(sce, title = "Expression frequency vs. mean expression level")
}
\seealso{
\code{\link[scater:plotRowData]{scater::plotRowData()}}, \code{\link[scuttle:addPerCellQCMetrics]{scuttle::addPerFeatureQCMetrics()}},
\code{\link[scuttle:perFeatureQCMetrics]{scuttle::perFeatureQCMetrics()}}, \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}
}
\author{
I-Hsuan Lin
}
