% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sc.R
\name{plotVariableFeature}
\alias{plotVariableFeature}
\title{Plot variance modelling results}
\usage{
plotVariableFeature(
  sce,
  var,
  hvg = NULL,
  top_n = 10,
  point_size = 2,
  text_size = 4,
  text_color = "black",
  trend_color = "gold",
  trend_size = 2,
  box.padding = 0.5,
  max.overlaps = Inf,
  seed = 12321,
  xlab = "Mean log-counts",
  ylab = "Variance of log-counts",
  title = NULL,
  theme_size = 18,
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{var}{A DataFrame returned by variance modelling functions such as
\code{modelGeneVar()}, \code{modelGeneVarByPoisson()}, etc.}

\item{hvg}{A character vector containing a set of highly variable genes
to highlight in red. Default is \code{NULL}.}

\item{top_n}{An integer scalar indicating the first \code{top_n} genes to label.
When \code{hvg = NULL}, the first \code{top_n} genes with the highest \code{mean} are
selected. When HVGs are provided, the first \code{top_n} HVGs are selected.
Default is 10.}

\item{point_size}{A numeric scalar indicating the size of the points.
Default is 2.}

\item{text_size}{A numeric scalar indicating the size of the label. This
is passed to \code{geom_text_repel()}. Default is 4.}

\item{text_color}{A string indicating the colour of the label. This is
passed to \code{geom_text_repel()}. Default is "black".}

\item{trend_color}{A string indicating the colour of the smoothed curve.
Default is "gold",}

\item{trend_size}{A numeric scalar indicating the size of the smoothed
curve. Default is 2.}

\item{box.padding}{A scalar indicating the amount of padding around
bounding box, as unit or number. This is passed to \code{geom_text_repel()}.
Default is 0.5.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
This is passed to \code{geom_text_repel()}. Default is Inf.}

\item{seed}{Random seed passed to \code{\link[=set.seed]{set.seed()}}. This is passed to
\code{geom_text_repel()}. Default is 12321.}

\item{xlab}{The title of the x-axis. Default is "Mean log-counts".}

\item{ylab}{The title of the y-axis. Default is "Variance of log-counts".}

\item{title}{Plot title. Default is \code{NULL}.}

\item{theme_size}{A numeric scalar indicating the base font size.
Default is 18.}

\item{...}{Other arguments passed on to \code{geom_text_repel()}.}
}
\value{
A \code{ggplot} object
}
\description{
This function requires a \code{SingleCellExperiment} object and the resulting
DataFrame after performing variance modelling, such as \code{modelGeneVar()}
from the \pkg{scran} package, and produces a mean-variance scatter plot.
}
\details{
In addition to the two required inputs, if \code{rowData(sce)} has \code{is_pass}
and/or \code{is_ambient} columns containing logical values (i.e. \code{TRUE} and
\code{FALSE}), denoting if a gene passed QC or if its expression is attributable
to ambient contamination, the resulting figure will show the genes in
different shapes. Also, genes that are HVGs are coloured in red when a gene
list is provided using the \code{hvg} argument.
}
\examples{
library(SingleCellExperiment)

# Load demo dataset
data(sce)

var <- metadata(sce)[["modelGeneVar"]]
hvg <- metadata(sce)[["HVG"]]

# View modelGeneVar result
plotVariableFeature(sce, var, top_n = 5, title = "modelGeneVar")

# Highlight HVGs
plotVariableFeature(sce, var, hvg, title = "modelGeneVar (highlight HVGs)")
}
\seealso{
\code{\link[scran:modelGeneVar]{scran::modelGeneVar()}}, \code{\link[scran:modelGeneVarByPoisson]{scran::modelGeneVarByPoisson()}},
\code{\link[scran:modelGeneVarWithSpikes]{scran::modelGeneVarWithSpikes()}}, \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}
}
\author{
I-Hsuan Lin
}
